third_party/blink/web_tests/fast/events/init-events-expected.txt

This tests the init functions for all the event DOM classes that have them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testInitEvent('', '"a", false, false').type is 'a'
PASS testInitEvent('', 'null, false, false').type is 'null'
PASS testInitEvent('', '"a", false, false').bubbles is false
PASS testInitEvent('', '"a", true, false').bubbles is true
PASS testInitEvent('', '"a", false, false').cancelable is false
PASS testInitEvent('', '"a", false, true').cancelable is true
PASS testInitEvent('Keyboard', '"a", false, false, window, "b", 1001, false, false, false, false, false').type is 'a'
PASS testInitEvent('Keyboard', 'null, false, false, window, "b", 1001, false, false, false, false, false').type is 'null'
PASS testInitEvent('Keyboard', '"a", false, false, window, "b", 1001, false, false, false, false, false').bubbles is false
PASS testInitEvent('Keyboard', '"a", true, false, window, "b", 1001, false, false, false, false, false').bubbles is true
PASS testInitEvent('Keyboard', '"a", false, false, window, "b", 1001, false, false, false, false, false').cancelable is false
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, false').cancelable is true
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, false').view is window
PASS testInitEvent('Keyboard', '"a", false, true, null, "b", 1001, false, false, false, false, false').view is null
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, false').location is 1001
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, false').ctrlKey is false
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, true, false, false, false, false').ctrlKey is true
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, false').altKey is false
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, true, false, false, false').altKey is true
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, false').shiftKey is false
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, true, false, false').shiftKey is true
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, false').metaKey is false
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, true, false').metaKey is true
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, true').detail is 0
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, false').keyCode is 0
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, false').charCode is 0
PASS testInitEvent('Keyboard', '"a", false, true, window, "b", 1001, false, false, false, false, false').which is 0
PASS testInitEvent('Message', '"a", false, false, "b", "c", "d", window, undefined').type is 'a'
PASS testInitEvent('Message', 'null, false, false, "b", "c", "d", window, undefined').type is 'null'
PASS testInitEvent('Message', '"a", false, false, "b", "c", "d", window, undefined').bubbles is false
PASS testInitEvent('Message', '"a", true, false, "b", "c", "d", window, undefined').bubbles is true
PASS testInitEvent('Message', '"a", false, false, "b", "c", "d", window, undefined').cancelable is false
PASS testInitEvent('Message', '"a", false, true, "b", "c", "d", window, undefined').cancelable is true
PASS testInitEvent('Message', '"a", false, false, "b", "c", "d", window, undefined').data is 'b'
PASS testInitEvent('Message', '"a", false, false, null, "c", "d", window, undefined').data is null
PASS testInitEvent('Message', '"a", false, false, "b", "c", "d", window, undefined').origin is 'c'
PASS testInitEvent('Message', '"a", false, false, "b", null, "d", window, undefined').origin is 'null'
PASS testInitEvent('Message', '"a", false, false, "b", "c", "d", window, undefined').lastEventId is 'd'
PASS testInitEvent('Message', '"a", false, false, "b", "c", null, window, undefined').lastEventId is 'null'
PASS testInitEvent('Message', '"a", false, false, "b", "c", "d", window, undefined').source is window
PASS testInitEvent('Message', '"a", false, false, "b", "c", "d", null, undefined').source is null
PASS testInitEvent('Message', '"a", false, false, "b", "c", "d", null, channelArray').ports[0] is channel.port1
PASS testInitEvent('Message', '"a", false, false, "b", "c", "d", null, channelArray').ports[1] is channel.port2
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').type is 'a'
PASS testInitEvent('Mouse', 'null, false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').type is 'null'
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').bubbles is false
PASS testInitEvent('Mouse', '"a", true, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').bubbles is true
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').cancelable is false
PASS testInitEvent('Mouse', '"a", false, true, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').cancelable is true
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').view is window
PASS testInitEvent('Mouse', '"a", false, false, null, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').view is null
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').detail is 1001
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').screenX is 1002
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').screenY is 1003
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').clientX is 1004
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').clientY is 1005
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').ctrlKey is false
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, true, false, false, false, 1006, null').ctrlKey is true
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').altKey is false
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, true, false, false, 1006, null').altKey is true
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').shiftKey is false
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, true, false, 1006, null').shiftKey is true
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').metaKey is false
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, true, 1006, null').metaKey is true
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').button is 1006
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').relatedTarget is null
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, document').relatedTarget is document
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').offsetX is 1004
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').offsetY is 1005
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').x is 1004
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').y is 1005
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').fromElement is null
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, document').fromElement is document
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').toElement is null
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').layerX is 1004
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').layerY is 1005
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').pageX is 1004
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').pageY is 1005
PASS testInitEvent('Mouse', '"a", false, false, window, 1001, 1002, 1003, 1004, 1005, false, false, false, false, 1006, null').which is 1007
PASS testInitEvent('Mutation', '"a", false, false, null, "b", "c", "d", 1001').type is 'a'
PASS testInitEvent('Mutation', 'null, false, false, null, "b", "c", "d", 1001').type is 'null'
PASS testInitEvent('Mutation', '"a", false, false, null, "b", "c", "d", 1001').bubbles is false
PASS testInitEvent('Mutation', '"a", true, false, null, "b", "c", "d", 1001').bubbles is true
PASS testInitEvent('Mutation', '"a", false, false, null, "b", "c", "d", 1001').cancelable is false
PASS testInitEvent('Mutation', '"a", false, true, null, "b", "c", "d", 1001').cancelable is true
PASS testInitEvent('Mutation', '"a", false, false, null, "b", "c", "d", 1001').relatedNode is null
PASS testInitEvent('Mutation', '"a", false, false, document, "b", "c", "d", 1001').relatedNode is document
PASS testInitEvent('Mutation', '"a", false, false, null, "b", "c", "d", 1001').prevValue is 'b'
PASS testInitEvent('Mutation', '"a", false, false, null, null, "c", "d", 1001').prevValue is 'null'
PASS testInitEvent('Mutation', '"a", false, false, null, "b", "c", "d", 1001').newValue is 'c'
PASS testInitEvent('Mutation', '"a", false, false, null, "b", null, "d", 1001').newValue is 'null'
PASS testInitEvent('Mutation', '"a", false, false, null, "b", "c", "d", 1001').attrName is 'd'
PASS testInitEvent('Mutation', '"a", false, false, null, "b", "c", null, 1001').attrName is 'null'
PASS testInitEvent('Mutation', '"a", false, false, null, "b", "c", "d", 1001').attrChange is 1001
PASS testInitEvent('Storage', '"a", false, false, "b", "c", "d", "e"').type is 'a'
PASS testInitEvent('Storage', 'null, false, false, "b", "c", "d", "e"').type is 'null'
PASS testInitEvent('Storage', '"a", false, false, "b", "c", "d", "e"').bubbles is false
PASS testInitEvent('Storage', '"a", true, false, "b", "c", "d", "e"').bubbles is true
PASS testInitEvent('Storage', '"a", false, false, "b", "c", "d", "e"').cancelable is false
PASS testInitEvent('Storage', '"a", false, true, "b", "c", "d", "e"').cancelable is true
PASS testInitEvent('Storage', '"a", false, false, "b", "c", "d", "e"').key is 'b'
PASS testInitEvent('Storage', '"a", false, false, null, "c", "d", "e"').key is null
PASS testInitEvent('Storage', '"a", false, false, "b", "c", "d", "e"').oldValue is 'c'
PASS testInitEvent('Storage', '"a", false, false, "b", null, "d", "e"').oldValue is null
PASS testInitEvent('Storage', '"a", false, false, "b", "c", "d", "e"').newValue is 'd'
PASS testInitEvent('Storage', '"a", false, false, "b", "c", null, "e"').newValue is null
PASS testInitEvent('Storage', '"a", false, false, "b", "c", "d", "e"').url is 'e'
PASS testInitEvent('Storage', '"a", false, false, "b", "c", "d", null').url is 'null'
PASS testInitEvent('Text', '"a", false, false, window, "b"').type is 'a'
PASS testInitEvent('Text', 'null, false, false, window, "b"').type is 'null'
PASS testInitEvent('Text', '"a", false, false, window, "b"').bubbles is false
PASS testInitEvent('Text', '"a", true, false, window, "b"').bubbles is true
PASS testInitEvent('Text', '"a", false, false, window, "b"').cancelable is false
PASS testInitEvent('Text', '"a", false, true, window, "b"').cancelable is true
PASS testInitEvent('Text', '"a", false, false, window, "b"').view is window
PASS testInitEvent('Text', '"a", false, false, null, "b"').view is null
PASS testInitEvent('Text', '"a", false, false, window, "b"').data is 'b'
PASS testInitEvent('Text', '"a", false, false, window, null').data is 'null'
PASS testInitEvent('Text', '"a", false, false, window, "b"').detail is 0
PASS testInitEvent('Text', '"a", false, false, window, "b"').which is 0
PASS testInitEvent('UI', '"a", false, false, window, 1001').type is 'a'
PASS testInitEvent('UI', 'null, false, false, window, 1001').type is 'null'
PASS testInitEvent('UI', '"a", false, false, window, 1001').bubbles is false
PASS testInitEvent('UI', '"a", true, false, window, 1001').bubbles is true
PASS testInitEvent('UI', '"a", false, false, window, 1001').cancelable is false
PASS testInitEvent('UI', '"a", false, true, window, 1001').cancelable is true
PASS testInitEvent('UI', '"a", false, false, window, 1001').view is window
PASS testInitEvent('UI', '"a", false, false, null, 1001').view is null
PASS testInitEvent('UI', '"a", false, false, window, 1001').detail is 1001
PASS testInitEvent('UI', '"a", false, false, window, 1001').which is 0
PASS successfullyParsed is true

TEST COMPLETE
